refactor: 클래스를 파일로 분리
Operand, Result, Operation를 별도 파일로 분리한다.
